refactor: リーダブルコードの原則に基づきコードを改善
