+++ tests/test_performance.py
@@ -1,10 +1,7 @@
-import sys
-import os
 import time
+
 import pytest
 
-# プロジェクトのルートディレクトリをsys.pathに追加
-sys.path.insert(0, os.path.abspath(os.path.join(os.path.dirname(__file__), "..")))
 from gacha import gacha_draw
